+++ app/src/main/java/jp/kawagh/kiando/Question.kt
@@ -1,8 +1,13 @@
 package jp.kawagh.kiando
 
+import androidx.room.Embedded
 import androidx.room.Entity
+import androidx.room.Junction
 import androidx.room.PrimaryKey
+import androidx.room.Relation
 import androidx.room.TypeConverter
+import jp.kawagh.kiando.models.QuestionTagCrossRef
+import jp.kawagh.kiando.models.Tag
 
 class Converters {
     // encode Move -> fromRow_fromCol_toRow_toCol_isPromote
@@ -51,6 +56,20 @@ data class Question(
         get() = SFENConverter().convertKomadaiFrom(komadaiSfen).second
 
 }
+
+data class QuestionWithTags(
+    @Embedded val question: Question,
+    @Relation(
+        parentColumn = "id",
+        entityColumn = "id",
+        associateBy = Junction(
+            QuestionTagCrossRef::class,
+            parentColumn = "question_id",
+            entityColumn = "tag_id",
+        )
+    )
+    val tags: List<Tag>
+)
 
 val initialSFEN = "lnsgkgsnl/1r5b1/ppppppppp/9/9/9/PPPPPPPPP/1B5R1/LNSGKGSNL"
 val sample3_SFEN = "lnsgkgsnl/1r5b1/pppppp1pp/6p2/9/2P6/PP1PPPPPP/1B5R1/LNSGKGSNL b - 1"