+++ app/src/main/java/jp/kawagh/kiando/QuestionsViewModel.kt
@@ -6,10 +6,13 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.github.doyaaaaaken.kotlincsv.dsl.csvReader
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import jp.kawagh.kiando.data.AppDatabase
 import jp.kawagh.kiando.data.Repository
+import jp.kawagh.kiando.models.QuestionTagCrossRef
+import jp.kawagh.kiando.models.Tag
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
@@ -18,7 +21,7 @@ import javax.inject.Inject
 @HiltViewModel
 class QuestionsViewModel @Inject constructor(
     private val repository: Repository,
-    @ApplicationContext context: Context
+    @ApplicationContext private val context: Context,
 ) :
     ViewModel() {
     private val db = AppDatabase.getInstance(context)
@@ -26,8 +29,9 @@ class QuestionsViewModel @Inject constructor(
 
     init {
         viewModelScope.launch() {
-            db.questionDao().getAll().collectLatest {
-                uiState = uiState.copy(questions = it)
+            db.questionDao().getQuestionsWithTags().collectLatest {
+                uiState = uiState.copy(questionsWithTags = it)
+
             }
         }
     }
@@ -63,8 +67,43 @@ class QuestionsViewModel @Inject constructor(
         }
 
     }
+
+    /**
+     * Records with negative id are predefined.
+     */
+    fun addSampleQuestionsAndTags() {
+        if (uiState.questionsWithTags.isEmpty()) {
+            viewModelScope.launch(Dispatchers.IO) {
+                db.tagDao().insert(Tag(id = -1, title = "sample"))
+                sampleQuestions.reversed().forEachIndexed { index, question ->
+                    val questionId = -(index + 1)
+                    db.questionDao().insert(question.copy(id = questionId))
+                    db.questionTagCrossRefDao()
+                        .insert(QuestionTagCrossRef(questionId = questionId, tagId = -1))
+                }
+            }
+        }
+    }
+
+    fun loadQuestionsFromAsset() {
+        val csvQuestionStream = context.resources.openRawResource(R.raw.questions)
+        val csvReader = csvReader { this.delimiter = ';' }
+        val csvQuestionContents = csvReader.readAllWithHeader(csvQuestionStream)
+        println(csvQuestionContents)
+        csvQuestionContents.forEach {
+            val id = it.getValue("id").toInt()
+            val description: String = it.getValue("description")
+            val answerMove: Move = Converters().toMove(it.getValue("answerMove"))
+            val sfen: String = it.getValue("sfen")
+            val komadaiSfen: String = it.getValue("komadaiSfen")
+            val parsedQuestion = Question(id, description, answerMove, sfen, komadaiSfen)
+            viewModelScope.launch(Dispatchers.IO) {
+                db.questionDao().insert(parsedQuestion)
+            }
+        }
+    }
 }
 
 data class QuestionsUiState(
-    val questions: List<Question> = emptyList()
+    val questionsWithTags: List<QuestionWithTags> = emptyList()
 )