+++ app/src/main/java/jp/kawagh/kiando/data/AppDatabase.kt
@@ -1,6 +1,7 @@
 package jp.kawagh.kiando.data
 
 import android.content.Context
+import androidx.room.AutoMigration
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
@@ -9,11 +10,19 @@ import androidx.room.migration.Migration
 import androidx.sqlite.db.SupportSQLiteDatabase
 import jp.kawagh.kiando.Converters
 import jp.kawagh.kiando.Question
+import jp.kawagh.kiando.models.QuestionTagCrossRef
+import jp.kawagh.kiando.models.Tag
 
-@Database(entities = [Question::class], version = 3, exportSchema = false)
+@Database(
+    entities = [Question::class, Tag::class, QuestionTagCrossRef::class],
+    autoMigrations = [AutoMigration(4, 5)],
+    version = 7
+)
 @TypeConverters(Converters::class)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun questionDao(): QuestionDao
+    abstract fun tagDao(): TagDao
+    abstract fun questionTagCrossRefDao(): QuestionTagCrossRefDao
 
     companion object {
         private var INSTANCE: AppDatabase? = null
@@ -23,7 +32,8 @@ abstract class AppDatabase : RoomDatabase() {
                     context, AppDatabase::class.java,
                     "database"
                 )
-                    .addMigrations(MIGRATION_2_3)
+                    .addMigrations(MIGRATION2to3, MIGRATION3to4, MIGRATION6to7)
+                    .fallbackToDestructiveMigrationFrom(5)
                     .build()
             }
             return INSTANCE as AppDatabase
@@ -31,10 +41,21 @@ abstract class AppDatabase : RoomDatabase() {
     }
 }
 
-val MIGRATION_2_3 = object : Migration(2, 3) {
+val MIGRATION2to3 = object : Migration(2, 3) {
     override fun migrate(database: SupportSQLiteDatabase) {
         database.execSQL(
             "ALTER TABLE questions ADD COLUMN tag_id INTEGER"
+        )
+    }
+}
+val MIGRATION3to4 = object : Migration(3, 4) {
+    override fun migrate(database: SupportSQLiteDatabase) {
+    }
+}
+val MIGRATION6to7 = object : Migration(6, 7) {
+    override fun migrate(database: SupportSQLiteDatabase) {
+        database.execSQL(
+            """ DROP TABLE question_movie_cross_ref """.trimIndent()
         )
     }
 }