+++ app/src/main/java/jp/kawagh/kiando/data/QuestionDao.kt
@@ -2,12 +2,17 @@ package jp.kawagh.kiando.data
 
 import androidx.room.*
 import jp.kawagh.kiando.Question
+import jp.kawagh.kiando.QuestionWithTags
 import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface QuestionDao {
     @Query("SELECT * FROM questions")
     fun getAll(): Flow<List<Question>>
+
+    @Transaction
+    @Query("SELECT * FROM questions")
+    fun getQuestionsWithTags(): Flow<List<QuestionWithTags>>
 
     @Query("SELECT * FROM questions WHERE id = :questionId")
     fun findById(questionId: Int): Question