+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # CHANGELOG
+
+## unreleased
+- [experimental] tag
+
+## v1.0.7 [2023-02-16]
+- add option to delete sample questions
+- add option to add sample questions
+
+## v1.0.6 [2023-02-12]
+- add feature to change question name
+- change ShogiBoard position to top
+- show Move to register
+
 ## v1.0.5 [2022-10-29]
 - display answer
 - rotate pieces on Komadai